+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 UNRELEASED
 ----------
 
+* `#90 <https://github.com/pytest-dev/pytest-reportlog/issues/90>`_: Fix jsonl output for pytest subtests (introduced in pytest 9.0).
 * Added official support for Python 3.12 and 3.13.
 * Dropped support for EOL Python 3.7 and Python 3.8.